Remove WENOPolynomial and integrate into WENOBase
WENOPolynomial only added the calculated values to the matrix A for
each row. This is now included in the Base class.

Further, no bound checking was possible for WENOPolynomial due to
the used scalar pointer. The new method uses possible bounds checking by using
the Foam::Matrix::operator[]() that uses bounds checking if compiled
with DFULLDEBUG flag.