Revert changes of commit: bd745dbcae7348930810f7671dce98f96ab29108
The intBasTrans value does change with the neighbour or owner face,
as the volume integral for the different cells can be different.

The face area however is the same, therefore here only one side is
stored.

To improve readability the Foam::Pair is used instead of Foam::List
to store the two sides. This shows then directly in the name that only
two entries are present.