[BugFix] Correct initial reference frame (#33)
The initial reference frame could have less than 3 edges or give a singular
Jacobi matrix which leads to a fatal error in the Foam::det() function.

Now all possible reference frames are checked and if the first possible one is
found it is selected. Hence, avoiding the second while loop.

This commit introduces the checkReferenceFrame() function.