Use one MPI_Waitall for collectData in WENOCoeff
For large scale simulations it appears that one
MPI_Waitall is better than using MPI_Wait for each request.
Also using MPI_Test in calcCoeff() for a real non-blocking communication
increased communication time for large number of processors (>2000)

Why this behavior appears is not clear and is against what would generally be
expected.